requirements.txt
openai
langchain
qdrant-client
streamlit
tqdm
python-dotenv
pandas
sentence-transformers
fastembed
